Updated and improved logic for the register route
