+++ frontend/app/routes/register/schemas.ts
@@ -1,0 +1,4 @@
+import {z} from "zod";
+
+
+
